TST: collect tests by method from test_api
